Move vec-cycle-wrapped test
